requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-environ==0.4.5
djangorestframework==3.10.2
ez-setup==0.9
idna==2.8
mws==0.8.6
python-flipkart==0.2.0
pytz==2019.2
requests==2.22.0
sanction==0.4.1
sqlparse==0.3.0
urllib3==1.25.3
